# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo API",
        "version": "1.0"
    },
    "paths": {
...
        "/areaCodes/findAll": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Find all area codes",
                "operationId": "areaCodes_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AreaCodesFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/cities/findAll": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Find all cities",
                "operationId": "cities_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/months/findAll": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Find all months",
                "operationId": "months_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/companies/delete/{companyId}": {
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete company",
                "operationId": "companyies_delete",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "description": "Company uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/departments/update/{gradeId}": {
            "put": {
                "tags": [
                    "Departments"
                ],
                "summary": "Update employee grade",
                "operationId": "departments_update",
                "parameters": [
                    {
                        "name": "gradeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Company",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepartmentSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/employees/delete/{employeeId}": {
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete employee",
                "operationId": "employees_delete",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "path",
                        "description": "Employee uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/jobtitlelists/preview": {
            "get": {
                "tags": [
                    "JobTitle Lists"
                ],
                "summary": "Get excel file to preview results",
                "operationId": "jobTitleLists_previewResults",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/jobtitlelists/reset": {
            "delete": {
                "tags": [
                    "JobTitle Lists"
                ],
                "summary": "Reset all statuses to restart operation",
                "operationId": "jobTitleLists_reset",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AreaCodesFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "phoneMask": {
                                    "type": "string"
                                },
                                "mobileMask": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "CommonFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "DepartmentSave": {
                "properties": {
                    "departmentId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "companyId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "departmentName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ObjectId": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}