# src/App/src/Handler/Common/AreaCodes/FindAllHandler.php
<?php

declare(strict_types=1);

namespace App\Handler\Common\AreaCodes;

use App\Model\CommonModel;
use Laminas\Diactoros\Response\JsonResponse;
use Psr\Http\Message\ResponseInterface;
use Psr\Http\Message\ServerRequestInterface;
use Psr\Http\Server\RequestHandlerInterface;

class FindAllHandler implements RequestHandlerInterface
{
    public function __construct(CommonModel $commonModel)
    {
        $this->commonModel = $commonModel;
    }

    /**
     * @OA\Get(
     *   path="/areaCodes/findAll",
     *   tags={"Common"},
     *   summary="Find all area codes",
     *   operationId="areaCodes_findAll",
     *   
     *   @OA\Response(
     *     response=200,
     *     description="Successful operation",
     *     @OA\JsonContent(ref="#/components/schemas/AreaCodesFindAll"),
     *   ),
     *   @OA\Response(
     *      response=404,
     *      description="No result found"
     *   )
     *)
     **/
    public function handle(ServerRequestInterface $request): ResponseInterface
    {
        $data = $this->commonModel->findAreaCodes();
        return new JsonResponse([
            'data' => $data
        ]);
    }

}